replace ModuleSpecsIndex with stub-based
