+++ src/main/kotlin/org/move/lang/core/stubs/impl/MvFileStub.kt
@@ -18,7 +18,7 @@ class MvFileStub(file: MoveFile?) : PsiFileStubImpl<MoveFile>(file) {
     override fun getType() = Type
 
     object Type : IStubFileElementType<MvFileStub>(MoveLanguage) {
-        private const val STUB_VERSION = 9
+        private const val STUB_VERSION = 10
 
         // Bump this number if Stub structure changes
         override fun getStubVersion(): Int = MoveParserDefinition.PARSER_VERSION + STUB_VERSION
